Refactor dashboard layout with persistent sidebar and settings sections
