+++ frontend/src/pages/tasks/TasksSidebar.tsx
@@ -3,13 +3,13 @@
 import type { Task, Workspace } from '@prisma/client'
 
 // Core
-import { Link, useNavigate } from 'react-router-dom'
+import { Link, useLocation, useNavigate } from 'react-router-dom'
 
 // Lib
 import { useConfirm } from '@frontend/hooks/useConfirm'
 
 // Redux
-import { dispatch } from '@frontend/framework/store'
+import { dispatch, useSelector } from '@frontend/framework/store'
 import { taskActions } from '@frontend/framework/redux/stores/tasks'
 
 // User interface
@@ -18,21 +18,31 @@ import { Button, Card, Tooltip } from '@heroui/react'
 // Misc
 import { getTaskViewUrl, UrlTree } from '@common/urls'
 import { deleteTask } from '@frontend/lib/routes/taskRoutes'
-import { DeleteIcon, PlusIcon } from '@frontend/common/IconNexus'
-
-type Props = {
-  tasks: Task[]
-  workspaces: Workspace[]
-  selectedTaskId?: string
-}
+import {
+  DeleteIcon,
+  PlusIcon,
+  SettingsIcon,
+} from '@frontend/common/IconNexus'
 
 function getWorkspaceLabel(workspaces: Workspace[], workspaceId: string) {
   const workspace = workspaces.find((entry) => entry.id === workspaceId)
   return workspace?.name || 'Unknown workspace'
 }
 
-export function TasksSidebar(props: Props) {
-  const { tasks, workspaces, selectedTaskId } = props
+function getSelectedTaskId(pathname: string) {
+  if (!pathname.startsWith('/tasks/')) {
+    return undefined
+  }
+
+  const pathParts = pathname.split('/')
+  return pathParts[2]
+}
+
+export function TasksSidebar() {
+  const tasks = useSelector((state) => state.tasks.items)
+  const workspaces = useSelector((state) => state.workspaces.items)
+  const location = useLocation()
+  const selectedTaskId = getSelectedTaskId(location.pathname)
   const confirm = useConfirm()
   const navigate = useNavigate()
 
@@ -68,59 +78,88 @@ export function TasksSidebar(props: Props) {
 
   return <aside className={`h-full w-72 shrink-0 border-r border-black/5 bg-white/70 p-4 backdrop-blur
   dark:border-white/10 dark:bg-slate-950/70`}>
-    <Button
-      as={Link}
-      to={UrlTree.tasksList}
-      className='relaxed w-full justify-start'
-      startContent={<PlusIcon />}
-      color='primary'
-    >
-      <span>New Task</span>
-    </Button>
-    <div className='compact text-xs uppercase tracking-wide opacity-50'>Recent Tasks</div>
-    <div className='relaxed space-y-2'>
-      {tasks.length === 0 && (
-        <Card className='p-3 dark:bg-slate-900/70'>
-          <div className='text-sm opacity-70'>No tasks yet.</div>
-        </Card>
-      )}
-      {tasks.map((task) => {
-        const isActive = task.id === selectedTaskId
-
-        return <Card
-          key={task.id}
-          className={`p-3 transition dark:bg-slate-900/60 ${
-            isActive
-              ? 'border border-black/10 bg-white dark:border-white/10 dark:bg-slate-900'
-              : 'bg-white/50'
-          }`}
+    <div className='flex h-full flex-col'>
+      <div>
+        <div className='relaxed text-xs uppercase tracking-wide opacity-50'>Navigation</div>
+        <div className='relaxed flex flex-col gap-2'>
+          <Button as={Link} to={UrlTree.tasksList} variant='light' className='w-full justify-start'>
+            <span>Tasks</span>
+          </Button>
+          <Button as={Link} to={UrlTree.workspacesList} variant='light' className='w-full justify-start'>
+            <span>Workspaces</span>
+          </Button>
+          <Button as={Link} to={UrlTree.llms} variant='light' className='w-full justify-start'>
+            <span>LLMs</span>
+          </Button>
+        </div>
+      </div>
+      <div className='relaxed'>
+        <Button
+          as={Link}
+          to={UrlTree.tasksList}
+          className='relaxed w-full justify-start'
+          startContent={<PlusIcon />}
+          color='primary'
         >
-          <div className='level'>
-            <Link to={getTaskViewUrl(task.id)} className='block min-w-0 flex-1'>
-              <div className='text-sm font-semibold truncate'>{task.name || 'Untitled task'}</div>
-              <div className='text-xs opacity-60 truncate'>{
-                getWorkspaceLabel(workspaces, task.workspaceId)
-              }</div>
-            </Link>
-            <Tooltip content='Delete Task'>
-              <div>
-                <Button
-                  isIconOnly
-                  variant='light'
-                  onPress={() => {
-                    handleDeleteTask(task)
-                  }}
-                  className='opacity-50 hover:opacity-100'
-                >
-                  <span className='icon'>
-                    <DeleteIcon />
-                  </span>
-                </Button>
+          <span>New Task</span>
+        </Button>
+        <div className='compact text-xs uppercase tracking-wide opacity-50'>Recent Tasks</div>
+        <div className='relaxed space-y-2'>
+          {tasks.length === 0 && (
+            <Card className='p-3 dark:bg-slate-900/70'>
+              <div className='text-sm opacity-70'>No tasks yet.</div>
+            </Card>
+          )}
+          {tasks.map((task) => {
+            const isActive = task.id === selectedTaskId
+
+            return <Card
+              key={task.id}
+              className={`p-3 transition dark:bg-slate-900/60 ${
+                isActive
+                  ? 'border border-black/10 bg-white dark:border-white/10 dark:bg-slate-900'
+                  : 'bg-white/50'
+              }`}
+            >
+              <div className='level'>
+                <Link to={getTaskViewUrl(task.id)} className='block min-w-0 flex-1'>
+                  <div className='text-sm font-semibold truncate'>{task.name || 'Untitled task'}</div>
+                  <div className='text-xs opacity-60 truncate'>
+                    {getWorkspaceLabel(workspaces, task.workspaceId)}
+                  </div>
+                </Link>
+                <Tooltip content='Delete Task'>
+                  <div>
+                    <Button
+                      isIconOnly
+                      variant='light'
+                      onPress={() => {
+                        handleDeleteTask(task)
+                      }}
+                      className='opacity-50 hover:opacity-100'
+                    >
+                      <span className='icon'>
+                        <DeleteIcon />
+                      </span>
+                    </Button>
+                  </div>
+                </Tooltip>
               </div>
-            </Tooltip>
-          </div>
-        </Card>
-      })}
+            </Card>
+          })}
+        </div>
+      </div>
+      <div className='mt-auto pt-4'>
+        <Button
+          as={Link}
+          to={UrlTree.settingsGeneral}
+          variant='light'
+          className='w-full justify-start'
+          startContent={<SettingsIcon />}
+        >
+          <span>Settings</span>
+        </Button>
+      </div>
     </div>
   </aside>
 }
